+++ demo/meeting/src/views/Meeting/index.tsx
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React from 'react';
-import { VideoTileGrid } from 'amazon-chime-sdk-component-library-react';
+import { VideoTileGrid, UserActivityProvider, UserActivityManager } from 'amazon-chime-sdk-component-library-react';
 
 import { StyledLayout, StyledContent } from './Styled';
 import NavigationControl from '../../containers/Navigation/NavigationControl';
@@ -16,13 +16,17 @@ const MeetingView = () => {
   const { showNavbar, showRoster } = useNavigation();
 
   return (
-    <StyledLayout showNav={showNavbar} showRoster={showRoster}>
-      <StyledContent>
-        <VideoTileGrid noRemoteVideoView={<MeetingDetails />} />
-        <MeetingControlsContainer />
-      </StyledContent>
-      <NavigationControl />
-    </StyledLayout>
+    <UserActivityProvider>
+      <StyledLayout showNav={showNavbar} showRoster={showRoster}>
+        <StyledContent>
+          <VideoTileGrid noRemoteVideoView={<MeetingDetails />} />
+          <UserActivityManager>
+            <MeetingControlsContainer />
+          </UserActivityManager>
+        </StyledContent>
+        <NavigationControl />
+      </StyledLayout>
+    </UserActivityProvider>
   );
 };